Big changes
Made the ability to add points to the wave, kinda to simulate player
movement. so when  camera moves sine wave adds points to direction its
moving and removes from other side. also started to work on making it so
amplitude and wave width can change while wave is moving but not
finsihed yet